mole/requirements.txt
django==2.2.28
dj-static==0.0.6
djangorestframework==3.13.1 
django-filter==2.4.0
pytz==2018.9
djangorestframework-gis==0.14
static==1.1.1
easy-thumbnails==2.6
django-extensions==2.1.6
django-silk==4.0.1
factory-boy==2.11.1
pandas==0.24.1
rest-pandas==1.0.0
ipdb==0.11
gunicorn==19.9.0
matplotlib==2.2.3
psycopg2-binary==2.7.7
pytest==4.5.0
xlrd==1.1.0
django-redis-cache==2.0.0
pulsar-client==2.9.1
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.8.0
debugpy
uritemplate==3.0.1
pyyaml==5.4.1
asteval==0.9.13
pydotplus==2.0.2
markdown==3.3.4